ROC-121, ROC-149 - Создание реализации обобщенного репозитория и UoW

See merge request pvt-q1-2018/rocket!46

+++ Rocket.BL.Common/Models/UserRoles/IPermission.cs
@@ -5,6 +5,6 @@
         /// <summary>
         /// id пермишена.
         /// </summary>
-        ushort PermissionId { get; set; }
+        int PermissionId { get; set; }
     }
 }
